// packages/datadog-api-client-v1/models/Creator.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Object describing the creator of the shared element.
*/
export declare class Creator {
    /**
    * Email of the creator.
    */
    'email'?: string;
    /**
    * Handle of the creator.
    */
    'handle'?: string;
    /**
    * Name of the creator.
    */
    'name'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): Creator;
    static serialize(data: Creator): {
        [key: string]: any;
    };
    constructor();
}